Refactoring of finders.
