+++ src/Knp/Bundle/KnpBundlesBundle/Tests/Finder/Fixtures/github-results.html
@@ -1,0 +1,97 @@
+<!DOCTYPE html>
+<html>
+  <head prefix="og: http://ogp.me/ns# fb: http://ogp.me/ns/fb# githubog: http://ogp.me/ns/fb/githubog#">
+    <meta charset='utf-8'>
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+        <title>Search · symfony2 bundle · GitHub</title>
+    <link rel="search" type="application/opensearchdescription+xml" href="/opensearch.xml" title="GitHub" />
+    <link rel="fluid-icon" href="https://github.com/fluidicon.png" title="GitHub" />
+    <link rel="shortcut icon" href="/favicon.ico" type="image/x-icon" />
+  </head>
+  <body class="logged_out    env-production " data-blob-contribs-enabled="yes">
+    <div id="wrapper">
+      <div id="header" class="true clearfix">
+        <div class="container clearfix">
+          <a class="site-logo" href="https://github.com/">
+          </a>
+            <ul class="top-nav logged_out">
+                <li class="pricing"><a href="https://github.com/plans">Signup and Pricing</a></li>
+                <li class="explore"><a href="https://github.com/explore">Explore GitHub</a></li>
+              <li class="features"><a href="https://github.com/features">Features</a></li>
+                <li class="blog"><a href="https://github.com/blog">Blog</a></li>
+              <li class="login"><a href="https://github.com/login?return_to=%2Fsearch%3Fq%3Dsymfony2%2Bbundle">Login</a></li>
+            </ul>
+        </div>
+      </div>
+      <div class="site clearfix">
+        <div id="site-container" class="container" data-pjax-container>
+          <div class="pagehead">
+            <h1>Search Results</h1>
+            <ul class="tabs" data-pjax>
+              <li><a href="/explore" highlight="explore_main">Explore</a></li>
+              <li><a href="/repositories" highlight="explore_repos">Repositories</a></li>
+              <li><a href="/languages" highlight="explore_languages">Languages</a></li>
+              <li><a href="/timeline" highlight="explore_timeline">Timeline</a></li>
+              <li><a href="/search" class="selected" highlight="code_search">Search</a></li>
+            </ul>
+        </div>
+        <div class="clear"></div>
+        <div id="code_search_results">
+            <table width="100%">
+              <tr><td width="60%" valign="top">
+                <div class="header">
+                  <div class="title">Repositories (2)</div>
+                  <div class="info">(0.0 seconds)</div>
+                </div>
+                <div class="result">
+                  <h2 class="title">
+                    <a href="/foo/bar">Foo bar</a>
+                    <span class="language">(PHP)</span>
+                  </h2>
+                  <div class="description">
+                    A Symfony2 Bundle for utilising the h5bp v1/2 html5 boilerplate code
+                  </div>
+                  <div class="details">
+                    140.6 KB <span>|</span>
+                    4 forks <span>|</span>
+                    20 watchers <span>|</span>
+                    last activity 6 months ago
+                  </div>
+                </div>
+                <div class="result">
+                  <h2 class="title">
+                    <a href="/foo/bar2">Foo bar2</a>
+                    <span class="language">(PHP)</span>
+                  </h2>
+                  <div class="description">
+                    A Symfony2 Bundle for utilising the h5bp v1/2 html5 boilerplate code
+                  </div>
+                  <div class="details">
+                    140.6 KB <span>|</span>
+                    4 forks <span>|</span>
+                    20 watchers <span>|</span>
+                    last activity 6 months ago
+                  </div>
+                </div>
+                <div class="result">
+                  <h2 class="title">
+                    <a href="https://nothing.com">nothing</a>
+                    <span class="language">(PHP)</span>
+                  </h2>
+                  <div class="description">
+                    A Symfony2 Bundle for utilising the h5bp v1/2 html5 boilerplate code
+                  </div>
+                  <div class="details">
+                    140.6 KB <span>|</span>
+                    4 forks <span>|</span>
+                    20 watchers <span>|</span>
+                    last activity 6 months ago
+                  </div>
+                </div>
+              </td></tr>
+            </table>
+      </div>
+    </div>
+    <span id='server_response_time' data-time='2.06957' data-host='fe12'></span>
+  </body>
+</html>
