+++ src/Knp/Bundle/KnpBundlesBundle/Tests/Finder/GithubTest.php
@@ -3,9 +3,9 @@
 namespace Knp\Bundle\KnpBundlesBundle\Tests\Finder;
 
 use Symfony\Component\DomCrawler\Crawler;
-use Knp\Bundle\KnpBundlesBundle\Finder\Google;
+use Knp\Bundle\KnpBundlesBundle\Finder\Github;
 
-class GoogleTest extends \PHPUnit_Framework_TestCase
+class GithubTest extends \PHPUnit_Framework_TestCase
 {
     /**
      * @test
@@ -22,24 +22,24 @@ class GoogleTest extends \PHPUnit_Framework_TestCase
             ->method('extract')
             ->will($this->returnCallback(function () use (&$callCounter) {
                 if ($callCounter) {
-                    return array('test' => 'https://github.com/KnpLabs/KnpBundles');
+                    return array('test' => '/KnpLabs/KnpBundles');
                 }
                 $callCounter++;
 
-                return array('test2' => 'https://github.com/l3l0/KnpBundles');
+                return array('test2' => '/l3l0/KnpBundles');
             }));
 
         $client = $this->getMock('Goutte\Client', array('request'));
         $client->expects($this->at(0))
             ->method('request')
-            ->with('GET', 'http://www.google.com/search?q=Symfony2')
+            ->with('GET', 'https://github.com/search?q=Symfony2&type=Repositories&language=PHP')
             ->will($this->returnValue($crawler));
         $client->expects($this->at(1))
             ->method('request')
-            ->with('GET', 'http://www.google.com/search?q=Symfony2&start=10')
+            ->with('GET', 'https://github.com/search?q=Symfony2&type=Repositories&language=PHP&start_value=30')
             ->will($this->returnValue($crawler));
 
-        $finder = new Google('Symfony2', 2, $client);
+        $finder = new Github('Symfony2', 2, $client);
         $repos = $finder->find();
 
         $this->assertEquals(array('l3l0/KnpBundles', 'KnpLabs/KnpBundles'), $repos);
@@ -49,7 +49,7 @@ class GoogleTest extends \PHPUnit_Framework_TestCase
      * @dataProvider getExtractPageUrlsData
      * @test
      */
-    public function shouldExtractPageUrlsFromGoogleHtml($node, $expected)
+    public function shouldExtractPageUrlsFromGithubHtml($node, $expected)
     {
         $callCounter = 0;
 
@@ -65,7 +65,7 @@ class GoogleTest extends \PHPUnit_Framework_TestCase
                 return new Crawler();
             }));
 
-        $finder = new Google('Symfony2', 3, $client);
+        $finder = new Github('Symfony2', 3, $client);
         $values = $finder->find();
 
         $this->assertEquals($expected, $values);
@@ -79,7 +79,7 @@ class GoogleTest extends \PHPUnit_Framework_TestCase
                 array()
             ),
             array(
-                file_get_contents(__DIR__.'/Fixtures/google-results.html'),
+                file_get_contents(__DIR__.'/Fixtures/github-results.html'),
                 array(
                     'foo/bar',
                     'foo/bar2'
@@ -94,7 +94,7 @@ class GoogleTest extends \PHPUnit_Framework_TestCase
      */
     public function shouldNotUseEmptyQuery()
     {
-        $finder = new Google('', 3);
+        $finder = new Github('', 3);
         $finder->find();
     }
 }
